Creates board, 6 attempts. must be 5 letters, update board, if correct letter in word and correct position char will be mark in green, if in wrong position it will be marked in yellow.
Game ends when attempts are maxed out or the word has been completed.

if there are double letters it will still show one of them in yellow if it is in the correct place already; also, even if there is only one, but if you guess a word with double letters it will be marked yellow because the logic will find it in the random word. need to find a fix for both of these things.

First push is a working terminal game with minor issues. Will fix in future pushes.